requirements.txt
fastapi
uvicorn
openai
python-dotenv
slowapi
pydantic
